Added different components
